+++ changelog.md
@@ -1,3 +1,8 @@
+## Changes in Ogre 3.4.3.0
+
+* Bump to Apache TinkerPop 3.4.3
+* Support new form of the `aggregate()` step that takes a `Scope`.
+
 ## Changes in Ogre 3.4.2.0
 
 * Bump to Apache TinkerPop 3.4.2